Refactor Call role API to better handle certificates
Changes:
Call::Role::CallState          -> Call::Role::State
Call::Role::CallLifeCycleState -> Call::Role::LifeCycleState
Call::Role::PhotoPtr           -> Call::Role::Photo
Call::Role::Direction2         -> Call::Role::Direction
Call::Role::PhoneNu            -> Call::Role::ContactMethod

Refs #66530 #66615

+++ src/contactproxymodel.cpp
@@ -366,7 +366,7 @@ QVariant ContactProxyModel::headerData(int section, Qt::Orientation orientation,
 bool ContactProxyModel::dropMimeData(const QMimeData *data, Qt::DropAction action, int row, int column, const QModelIndex &parent)
 {
    Q_UNUSED( action )
-   setData(parent,-1,Call::Role::DropState);
+   setData(parent,-1,static_cast<int>(Call::Role::DropState));
    if (data->hasFormat(RingMimes::CALLID)) {
       const QByteArray encodedCallId = data->data( RingMimes::CALLID    );
       const QModelIndex targetIdx    = index   ( row,column,parent );
